+++ dbs/src/dbsfmhrpt.f
@@ -143,7 +143,7 @@ C     BIND SQL STATEMENT PARAMETERS TO FORTRAN VARIABLES
      -           SQL_NULL_PTR)
 
       DO I=1,VARDIM
-        ColNumber=INT(ColNumber+1)
+        ColNumber=ColNumber+1
         iRet = fvsSQLBindParameter(StmtHndlOut,ColNumber,
      -    SQL_PARAM_INPUT,
      -    SQL_F_DOUBLE, SQL_DOUBLE,INT(15,SQLUINTEGER_KIND),
